Found the proper way to use a unique_ptr<Lexer> to initialize a Parser.
I missed the whole point of using unique_ptr<Lexer> in the previous
commit. unique_ptr can be used to initialize another object with a
member field of unique_ptr.
What I had to do was simply to "std::move" the original pointer to
whatever destination (function, constructor), because the whole purpose
of unique_ptr is to take ownership of a particular heap memory and pass
it around safely if necessary.